+++ applications/fishing-map/src/features/search/Search.tsx
@@ -8,7 +8,7 @@ import IconButton from '@globalfishingwatch/ui-components/dist/icon-button'
 import InputText from '@globalfishingwatch/ui-components/dist/input-text'
 import Spinner from '@globalfishingwatch/ui-components/dist/spinner'
 import useDebounce from '@globalfishingwatch/react-hooks/dist/use-debounce'
-import { Choice, Icon } from '@globalfishingwatch/ui-components'
+import { Button, Choice, Icon } from '@globalfishingwatch/ui-components'
 import { ChoiceOption } from '@globalfishingwatch/ui-components/dist/choice'
 import { useLocationConnect } from 'routes/routes.hook'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
@@ -45,7 +45,7 @@ function Search() {
   const { t } = useTranslation()
   const dispatch = useDispatch()
   const urlQuery = useSelector(selectSearchQuery)
-  const { upsertDataviewInstance } = useDataviewInstancesConnect()
+  const { addNewDataviewInstances } = useDataviewInstancesConnect()
   const [searchQuery, setSearchQuery] = useState((urlQuery || '') as string)
   const { searchFilters } = useSearchFiltersConnect()
   const { searchPagination, searchSuggestion, searchSuggestionClicked } = useSearchConnect()
@@ -57,6 +57,7 @@ function Search() {
   const searchStatus = useSelector(selectSearchStatus)
   const hasSearchFilters = checkSearchFiltersEnabled(searchFilters)
   const vesselDataviews = useSelector(selectVesselsDataviews)
+  const [vesselsSelected, setVesselsSelected] = useState<VesselWithDatasets[]>([])
 
   const searchOptions = [
     {
@@ -152,14 +153,27 @@ function Search() {
     }
   }
 
-  const onSelectionChange = (selection: VesselWithDatasets | null) => {
-    if (selection && selection.dataset && selection.trackDatasetId) {
-      const vesselDataviewInstance = getVesselDataviewInstance(selection, {
-        trackDatasetId: selection.trackDatasetId as string,
-        infoDatasetId: selection.dataset.id,
-      })
-      upsertDataviewInstance(vesselDataviewInstance)
+  const onSelect = (selection: VesselWithDatasets | null) => {
+    if (!selection) return
+    if (vesselsSelected.includes(selection)) {
+      setVesselsSelected(vesselsSelected.filter((vessel) => vessel !== selection))
+      return
     }
+    if (selection && selection.dataset && selection.trackDatasetId) {
+      setVesselsSelected([...vesselsSelected, selection])
+    }
+  }
+
+  const onConfirmSelection = () => {
+    const instances = vesselsSelected.map((vessel) => {
+      const vesselDataviewInstance = getVesselDataviewInstance(vessel, {
+        trackDatasetId: vessel.trackDatasetId as string,
+        infoDatasetId: vessel.dataset.id,
+      })
+      return vesselDataviewInstance
+    })
+    addNewDataviewInstances(instances)
+    onCloseClick()
   }
 
   const hasMoreResults =
@@ -184,7 +198,7 @@ function Search() {
   }
 
   return (
-    <Downshift onChange={onSelectionChange} itemToString={(item) => (item ? item.shipname : '')}>
+    <Downshift onSelect={onSelect} itemToString={(item) => (item ? item.shipname : '')}>
       {({ getInputProps, getItemProps, getMenuProps, highlightedIndex, selectedItem }) => (
         <div className={styles.search}>
           <div className={styles.header}>
@@ -259,12 +273,14 @@ function Search() {
                     const isInWorkspace = vesselDataviews?.some(
                       (vessel) => vessel.id === `${VESSEL_LAYER_PREFIX}${id}`
                     )
+                    const isSelected = vesselsSelected?.some((vessel) => vessel.id === id)
                     return (
                       <li
                         {...getItemProps({ item: entry, index })}
                         className={cx(styles.searchResult, {
                           [styles.highlighted]: highlightedIndex === index,
-                          [styles.selected]: isInWorkspace,
+                          [styles.inWorkspace]: isInWorkspace,
+                          [styles.selected]: isSelected,
                         })}
                         key={id}
                       >
@@ -326,10 +342,19 @@ function Search() {
                               </div>
                             )}
                           </div>
+                          {isSelected && (
+                            <span className={styles.alreadyAddedMsg}>
+                              <Icon icon="tick" />
+                              {t('search.vesselSelected', 'Vessel selected')}
+                            </span>
+                          )}
                           {isInWorkspace && (
                             <span className={styles.alreadyAddedMsg}>
                               <Icon icon="tick" />
-                              {t('search.vesselAlreadyInWorksace', 'Vessel already in workspace')}
+                              {t(
+                                'search.vesselAlreadyInWorkspace',
+                                'This vessel is already in your workspace'
+                              )}
                             </span>
                           )}
                         </Fragment>
@@ -354,6 +379,23 @@ function Search() {
             ) : (
               <SearchNotAllowed />
             )}
+          </div>
+          <div className={cx(styles.footer, { [styles.hidden]: vesselsSelected.length === 0 })}>
+            {vesselsSelected.length > 1 && (
+              <Button
+                disabled
+                type="secondary"
+                tooltip={t('common.comingSoon', 'Coming Soon')}
+                className={styles.footerAction}
+              >
+                See as fleet
+              </Button>
+            )}
+            <Button className={styles.footerAction} onClick={onConfirmSelection}>
+              {vesselsSelected.length > 1
+                ? t('search.seeVessels', 'See vessels')
+                : t('search.seeVessel', 'See vessel')}
+            </Button>
           </div>
         </div>
       )}